Guard every CrashC macro
Since CrashC heavily relies on macros in order to work properly
we must make sure that no macro used by CrashC is overridden.
Due to this, for every essential macro we check for its definition
and throw a compilation error. This may seem a bit tragic but it
assures that the framework does not behave in unknown ways which
would lead to painful debugging.
We also prepend the prefix "CT_" to every macro that is not directly
exposed to the final user, in order to try to avoid name clashes as
much as possible.